EID-1603 Move EidasValidatorFactory to saml-lib
The `EidasValidatorFactory` is used in the hub to validate the
signatures of SAML Responses and assertions from countries. We are going
to need to use this logic in the VSP and MSA as, due to countries not
signing assertions, the hub will be passing countries responses on to
the VSP/MSA without altering them.

The `EidasValidatorFactory` uses an `eidasMetadataResolverRepository`
which is responsible for fetching the countries trust anchors and
storing them in a trust store. The VSP and MSA will need to be
configured to do this.
